requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
aiosqlite==0.19.0
python-gitlab==3.15.0
PyYAML==6.0.2
jinja2==3.1.2
pillow==10.0.1
playwright==1.39.0